pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cherrytree"
dynamic = ["version"]
description = "🍒 Intelligent AI-assisted release management and cherry-picking for Apache Superset"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Maxime Beauchemin", email = "maximebeauchemin@gmail.com" },
]
keywords = [
    "release-management",
    "cherry-picking",
    "ci-cd",
    "git",
    "github",
    "superset",
    "automation",
    "ai-assisted"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=10.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "pygithub>=1.58.0",
    "packaging>=21.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
]
all = [
    "cherrytree[ai]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.9",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/apache-superset/cherrytree"
Repository = "https://github.com/apache-superset/cherrytree.git"
"Bug Tracker" = "https://github.com/apache-superset/cherrytree/issues"
Changelog = "https://github.com/apache-superset/cherrytree/blob/main/CHANGELOG.md"

[project.scripts]
cherrytree = "cherrytree.cli:main"
ct = "cherrytree.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["cherrytree"]

[tool.hatch.version]
path = "cherrytree/__init__.py"

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=cherrytree --cov-report=term-missing --cov-report=xml"
